Downgraded @openzepplin to support solidity v0.6.6

+++ contracts/controllers/SummerTimeVault.sol
@@ -60,7 +60,7 @@ contract SummmerTimeVault is Ownable, VaultCollateralConfig, UserVault {
         string collateralDisplayName,
         address token0Address,
         address token1Address,
-        uint minimumDebtCollateralRatio,
+        uint256 minimumDebtCollateralRatio,
         // address collateralAddress,
         // address stakingAddress,
         // address strategyAddress,
@@ -109,7 +109,9 @@ contract SummmerTimeVault is Ownable, VaultCollateralConfig, UserVault {
             token0: token0Address,
             token1: token1Address,
             priceOracleAddress: priceOracleAddress,
-            minimumDebtCollateralRatio: minimumDebtCollateralRatio || 50 * 10**17, // default: 1.5
+            minimumDebtCollateralRatio: minimumDebtCollateralRatio ||
+                50 *
+                10**17, // default: 1.5
             maxCollateralAmountAccepted: 1000 * 10**18
         });
